Trait rogue_dicekit::Dice

pub trait Dice: Rng {
    fn roll_target(
        &mut self, 
        count: usize, 
        sides: usize, 
        target: usize
    ) -> usize { ... }

    fn roll_sum(&mut self, count: usize, pips: usize) -> usize { ... }

    fn roll_pct(&mut self) -> f32 { ... }
}

Dice Extension to Rng

Provided methods

fn roll_target(&mut self, count: usize, sides: usize, target: usize) -> usize

Roll count dice with sides and count the total above target.

fn roll_sum(&mut self, count: usize, pips: usize) -> usize

Roll count dice with pips and sum the total pips.

fn roll_pct(&mut self) -> f32

Roll percentile dice and return the value between 0 and 100.